requirements.txt
azure-functions==1.18.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0

